sched/fair: always used unsigned long for utilization
Utilization and capacity are tracked as unsigned long, however some
functions using them returns an int which is ultimately assigned back to
unsigned long variables.

Since there is not scope on using a different and signed type, this
consolidate the signature of functions returning utilization to always
use the native type.
As well as improving code consistency this is expected also benefits
code paths where utilizations should be clamped by voiding further type
conversions or ugly type casts.

+++ kernel/sched/fair.c
@@ -5343,7 +5343,7 @@ struct energy_env {
 	} cap;
 };
 
-static int cpu_util_wake(int cpu, struct task_struct *p);
+static unsigned long cpu_util_wake(int cpu, struct task_struct *p);
 
 /*
  * __cpu_norm_util() returns the cpu util relative to a specific capacity,
@@ -6405,7 +6405,7 @@ done:
  * rq->curr. For that case we should return cpu util with contributions from
  * currently running task p removed.
  */
-static int cpu_util_wake(int cpu, struct task_struct *p)
+static unsigned long cpu_util_wake(int cpu, struct task_struct *p)
 {
 	unsigned long util, capacity;
